[refactor] controller & domain
